moved test module submissions dir to asset_groups
